Moved ComplementaryFormsCcsRecordAggregationHelper to the own file
